Backend/flask-app/requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-Migrate==3.1.0
Flask-Mail==0.9.1
flask-accept==0.0.6
flask-bcrypt==1.0.1
psycopg2-binary==2.9.3
SQLAlchemy==1.4.39
gunicorn==20.1.0
Werkzeug==2.1.2
pyjwt==2.4.0
redis==4.3.3
celery==5.2.7
coverage==6.4.1
alembic==1.8.0
email-validator==1.2.1
